Using test only in Model Compiler
